refactor(theme): move MeshCore theme into shared commonMain (multiplatform fonts)
Extract the MeshCore palette, shapes, Dimens, Section/SectionStates and branded
typography into :meshcore-components commonMain as the single source of truth,
with a multiplatform font setup (expect/actual): Android keeps the downloadable
Google Fonts provider (unchanged), desktop loads bundled .ttf. Adds a pure
commonMain MeshcoreTheme() for the shared composables + desktop parity render.

The app's theme files now re-export from components (typealias for Dimens/
Section/SectionStates; imports for colors/shapes/typography), so call-sites are
unchanged and the app's MeshcoreTheme keeps its Android dynamic-color +
edge-to-edge handling.

Verified: components desktop + android compile, and :app:compileDebugKotlin green.
